[Feature-10629][Improvement] Support multi cluster environments - namespace use cluster (#10650)
* server code

* namespace ui

* fix dao test

* fix test

* fix

* docs and image update

+++ dolphinscheduler-api/src/main/java/org/apache/dolphinscheduler/api/service/impl/ClusterServiceImpl.java
@@ -19,15 +19,20 @@ package org.apache.dolphinscheduler.api.service.impl;
 
 import org.apache.dolphinscheduler.api.dto.ClusterDto;
 import org.apache.dolphinscheduler.api.enums.Status;
+import org.apache.dolphinscheduler.api.k8s.K8sManager;
 import org.apache.dolphinscheduler.api.service.ClusterService;
 import org.apache.dolphinscheduler.api.utils.PageInfo;
 import org.apache.dolphinscheduler.api.utils.Result;
 import org.apache.dolphinscheduler.common.Constants;
+import org.apache.dolphinscheduler.common.utils.ClusterConfUtils;
 import org.apache.dolphinscheduler.common.utils.CodeGenerateUtils;
 import org.apache.dolphinscheduler.common.utils.CodeGenerateUtils.CodeGenerateException;
 import org.apache.dolphinscheduler.dao.entity.Cluster;
+import org.apache.dolphinscheduler.dao.entity.K8sNamespace;
 import org.apache.dolphinscheduler.dao.entity.User;
 import org.apache.dolphinscheduler.dao.mapper.ClusterMapper;
+import org.apache.dolphinscheduler.dao.mapper.K8sNamespaceMapper;
+import org.apache.dolphinscheduler.remote.exceptions.RemotingException;
 
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang.StringUtils;
@@ -46,6 +51,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 
@@ -60,6 +66,11 @@ public class ClusterServiceImpl extends BaseServiceImpl implements ClusterServic
     @Autowired
     private ClusterMapper clusterMapper;
 
+    @Autowired
+    private K8sManager k8sManager;
+
+    @Autowired
+    private K8sNamespaceMapper k8sNamespaceMapper;
     /**
      * create cluster
      *
@@ -237,6 +248,14 @@ public class ClusterServiceImpl extends BaseServiceImpl implements ClusterServic
             return result;
         }
 
+        Integer relatedNamespaceNumber = k8sNamespaceMapper
+            .selectCount(new QueryWrapper<K8sNamespace>().lambda().eq(K8sNamespace::getClusterCode, code));
+
+        if (relatedNamespaceNumber > 0) {
+            putMsg(result, Status.DELETE_CLUSTER_RELATED_NAMESPACE_EXISTS);
+            return result;
+        }
+
         int delete = clusterMapper.deleteByCode(code);
         if (delete > 0) {
             putMsg(result, Status.SUCCESS);
@@ -279,6 +298,16 @@ public class ClusterServiceImpl extends BaseServiceImpl implements ClusterServic
         if (clusterExist == null) {
             putMsg(result, Status.CLUSTER_NOT_EXISTS, name);
             return result;
+        }
+
+        if (!Constants.K8S_LOCAL_TEST_CLUSTER_CODE.equals(clusterExist.getCode())
+            && !config.equals(ClusterConfUtils.getK8sConfig(clusterExist.getConfig()))) {
+            try {
+                k8sManager.getAndUpdateK8sClient(code, true);
+            } catch (RemotingException e) {
+                putMsg(result, Status.K8S_CLIENT_OPS_ERROR, name);
+                return result;
+            }
         }
 
         //update cluster
